add DamageableSystem.SS220.cs
